fixing createReadStream() and organizing code
