+++ src/util.js
@@ -4,17 +4,54 @@ module.exports = {
   matchingNibbleLength: matchingNibbleLength,
   callTogether: callTogether,
   asyncFirstSeries: asyncFirstSeries,
-  doKeysMatch: doKeysMatch
+  doKeysMatch: doKeysMatch,
+  nibblesToBuffer: nibblesToBuffer,
+  stringToNibbles: stringToNibbles,
+}
+/**
+ * Converts a string OR a buffer to a nibble array.
+ * @method stringToNibbles
+ * @param {Buffer| String} key
+ * @private
+ */
+function stringToNibbles (key) {
+  const bkey = new Buffer(key)
+  let nibbles = []
+
+  for (let i = 0; i < bkey.length; i++) {
+    let q = i * 2
+    nibbles[q] = bkey[i] >> 4
+    ++q
+    nibbles[q] = bkey[i] % 16
+  }
+
+  return nibbles
 }
 
 /**
- * Returns the number of in order matching nibbles of two give nibble arrayes
+ * Converts a nibble array into a buffer.
+ * @method nibblesToBuffer
+ * @param {Array} Nibble array
+ * @private
+ */
+function nibblesToBuffer (arr) {
+  let buf = new Buffer(arr.length / 2)
+  for (let i = 0; i < buf.length; i++) {
+    let q = i * 2
+    buf[i] = (arr[q] << 4) + arr[++q]
+  }
+  return buf
+}
+
+/**
+ * Returns the number of in order matching nibbles of two give nibble arrays.
  * @method matchingNibbleLength
  * @param {Array} nib1
  * @param {Array} nib2
+ * @private
  */
 function matchingNibbleLength (nib1, nib2) {
-  var i = 0
+  let i = 0
   while (nib1[i] === nib2[i] && nib1.length > i) {
     i++
   }
@@ -22,10 +59,12 @@ function matchingNibbleLength (nib1, nib2) {
 }
 
 /**
- * Compare two 'nibble array' keys
+ * Compare two nibble array keys.
+ * @param {Array} keyA
+ * @param {Array} keyB
  */
 function doKeysMatch (keyA, keyB) {
-  var length = matchingNibbleLength(keyA, keyB)
+  const length = matchingNibbleLength(keyA, keyB)
   return length === keyA.length && length === keyB.length
 }
